Support for two-sticks, one controller
port: implements analog movement for pc controllers

port: Implement menu movement with the C stick

port: Implement analog leaning

port: Implement 1.2 + 1 analog zooming

port: pd input lib now handles 32 buttons

port: modern crouch implementation

port: implement analog slayer airbraking

port: xbla farsight strafing

Map crouch buttons

- map cycle crouch to L3
- sensible 1964-style crouch bindings

+++ src/game/debug2.c
@@ -538,7 +538,7 @@ void debug0f11944cnb(void) // not called
 }
 #endif
 
-bool debugProcessInput(s8 stickx, s8 sticky, u16 buttons, u16 buttonsthisframe)
+bool debugProcessInput(s8 stickx, s8 sticky, u32 buttons, u32 buttonsthisframe)
 {
 #ifdef DEBUG
 	s32 i;
